Move CreateOracleMessage and UpdateOracleMessage to public (to fix meaningful warning by vitepress)
