+++ src/oracleFactory/updateOracle.ts
@@ -4,11 +4,10 @@ import * as ipfs from '../services/ipfs/index.js';
 import {
   IExecConsumer,
   TaskExecutionMessage,
-  UpdateOracleMessage,
   UpdateOracleOptions,
   UpdateOracleParams,
 } from '../types/internal-types.js';
-import { ParamSet } from '../types/public-types.js';
+import { ParamSet, UpdateOracleMessage } from '../types/public-types.js';
 import { ValidationError, WorkflowError } from '../utils/errors.js';
 import { formatParamsJson } from '../utils/format.js';
 import { Observable, SafeObserver } from '../utils/reactive.js';
