+++ src/types/public-types.ts
@@ -1,6 +1,12 @@
 import { AbstractProvider } from 'ethers';
 import { EnhancedWallet } from 'iexec';
 import { IExecConfigOptions } from 'iexec/IExecConfig';
+import {
+  PublishedApporder,
+  PublishedDatasetorder,
+  PublishedRequestorder,
+  PublishedWorkerpoolorder,
+} from 'iexec/IExecOrderbookModule';
 
 /**
  * Ethereum address.
@@ -108,6 +114,132 @@ export type ParamSetCID = string;
  * The OracleID is unique for each ParamSet.
  */
 export type OracleID = string;
+
+/**
+ * Messages that will be sent by the observer in the process of creating an oracle
+ */
+interface DeployDatasetMessage {
+  message: 'DATASET_DEPLOYMENT_SUCCESS';
+  address: AddressOrENS;
+}
+
+interface CreateParamSetMessage {
+  message: 'PARAM_SET_CREATED';
+  paramSet: ParamSet;
+}
+
+interface ComputeOracleIDMessage {
+  message: 'ORACLE_ID_COMPUTED';
+  oracleId: string;
+}
+
+interface UploadParamSetMessage {
+  message: 'PARAM_SET_UPLOADED';
+  cid: string;
+  multiaddr: string;
+}
+
+export type CreateOracleMessage =
+  | DeployDatasetMessage
+  | CreateParamSetMessage
+  | ComputeOracleIDMessage
+  | UploadParamSetMessage;
+
+/**
+ * Messages that will be sent by the observer in the process of updating an oracle
+ */
+interface EnsureParamsMessage {
+  message: 'ENSURE_PARAMS';
+}
+
+interface EnsureParamsUploadMessage {
+  message: 'ENSURE_PARAMS_UPLOAD';
+}
+
+interface EnsureParamsSuccessMessage {
+  message: 'ENSURE_PARAMS_SUCCESS';
+  paramSet: ParamSet;
+  cid: string;
+}
+
+interface FetchAppOrderMessage {
+  message: 'FETCH_APP_ORDER';
+}
+
+interface FetchAppOrderSuccessMessage {
+  message: 'FETCH_APP_ORDER_SUCCESS';
+  order: PublishedApporder;
+}
+
+interface FetchDatasetOrderMessage {
+  message: 'FETCH_DATASET_ORDER';
+}
+
+interface FetchDatasetOrderSuccessMessage {
+  message: 'FETCH_DATASET_ORDER_SUCCESS';
+  order: PublishedApporder;
+}
+
+interface FetchWorkerpoolOrderMessage {
+  message: 'FETCH_WORKERPOOL_ORDER';
+}
+
+interface FetchWorkerpoolOrderSuccessMessage {
+  message: 'FETCH_WORKERPOOL_ORDER_SUCCESS';
+  order: PublishedApporder;
+}
+
+interface RequestOrderSignatureSignRequestMessage {
+  message: 'REQUEST_ORDER_SIGNATURE_SIGN_REQUEST';
+  order: PublishedApporder;
+}
+
+interface RequestOrderSignatureSuccessMessage {
+  message: 'REQUEST_ORDER_SIGNATURE_SUCCESS';
+  order: PublishedApporder;
+}
+
+interface MatchOrdersSignTxRequestMessage {
+  message: 'MATCH_ORDERS_SIGN_TX_REQUEST';
+  apporder: PublishedApporder;
+  datasetorder: PublishedDatasetorder;
+  workerpoolorder: PublishedWorkerpoolorder;
+  requestorder: PublishedRequestorder;
+}
+
+interface MatchOrdersSuccessMessage {
+  message: 'MATCH_ORDERS_SUCCESS';
+  dealid: string;
+  txHash: string;
+}
+
+interface TaskUpdatedMessage {
+  message: 'TASK_UPDATED';
+  dealid: string;
+  taskid: string;
+  status: string;
+}
+
+interface UpdateTaskCompletedMessage {
+  message: 'UPDATE_TASK_COMPLETED';
+}
+
+export type UpdateOracleMessage =
+  | EnsureParamsMessage
+  | EnsureParamsUploadMessage
+  | EnsureParamsSuccessMessage
+  | FetchAppOrderMessage
+  | FetchAppOrderSuccessMessage
+  | FetchDatasetOrderMessage
+  | FetchDatasetOrderSuccessMessage
+  | FetchWorkerpoolOrderMessage
+  | FetchWorkerpoolOrderSuccessMessage
+  | RequestOrderSignatureSignRequestMessage
+  | RequestOrderSignatureSuccessMessage
+  | MatchOrdersSignTxRequestMessage
+  | MatchOrdersSuccessMessage
+  | TaskUpdatedMessage
+  | UpdateTaskCompletedMessage;
 
 /**
  * Parameters for reading data from an oracle.
